% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convpoibin-package.R
\docType{package}
\name{convpoibin-package}
\alias{convpoibin-package}
\alias{convpoibin}
\title{Compute the Poisson binomial distribution via convolution}
\description{
The cdf and pmf for the Poisson binomial distribution computed using
convolution methods
}
\details{
\tabular{ll}{ Package: \tab convpoibin\cr Type: \tab Package\cr Version:
\tab 1.0\cr Date: \tab 2017-10-3\cr License: \tab GPL-2 \cr LazyLoad: \tab
yes\cr }

Two exact and efficient convolution based methods for computing the Poisson
binomial cdf and pmf.
}
\examples{

  p=runif(100)
  ppoibin(p,method='direct')
  ppoibin(p,method='fft_tree',M=2)
  dpoibin(p,method='direct')
  dpoibin(p,method='fft_tree',M=2)
  
  p=runif(1000)
  ppoibin(p,method='direct')
  ppoibin(p,method='fft_tree',M=4)
  dpoibin(p,method='direct')
  dpoibin(p,method='fft_tree',M=4)

}
\author{
William Biscarri

Maintainer: William Biscarri <wbiscar2@illinois.edu>
}
\keyword{package}
