% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dpoibin}
\alias{dpoibin}
\alias{ppoibin}
\alias{ppoibin}
\title{Compute the Poisson binomial distribution via convolution}
\usage{
dpoibin(p, method = "direct", M = 2)

ppoibin(p, method = "direct", M = 2)
}
\arguments{
\item{p}{The vector of success probabilities}

\item{method}{"direct" for the direct convolution method, "fft_tree" for
the binary tree fft convolution method.}

\item{M}{The number of splits for the tree used in the fft_tree
convolution.}
}
\value{
Returns the entire cdf or pmf
}
\description{
The cdf and pmf for the Poisson binomial distribution computed using
convolution methods
}
\details{
See the reference for computational details.
}
\examples{

  p=runif(100)
  ppoibin(p,method='direct',M=2)
  ppoibin(p,method='fft_tree',M=2)
  dpoibin(p,method='direct',M=2)
  dpoibin(p,method='fft_tree',M=2)
  
  p=runif(1000)
  ppoibin(p,method='direct',M=2)
  ppoibin(p,method='fft_tree',M=4)
  dpoibin(p,method='direct',M=2)
  dpoibin(p,method='fft_tree',M=4)

}
\references{
test test test
}
\author{
William Biscarri
}
